Created CommunicationBanner and using custom GroupBoxStyle
